% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{coin_toss}
\alias{coin_toss}
\title{Description pending}
\usage{
coin_toss(lose, win, players, rounds)
}
\arguments{
\item{lose}{percentage of wealth decrease if you lose
you lose to 0.6 of your current wealth}

\item{win}{percentage of wealth increase if you win
you win 1.5 of your current wealth}

\item{players}{number of players}

\item{rounds}{gamble rounds or repetitions}
}
\description{
Description pending
}
