% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{coin_toss}
\alias{coin_toss}
\title{Creating random sample}
\usage{
coin_toss(lose, win, players, rounds)
}
\arguments{
\item{lose}{percentage of wealth decrease if you lose
you lose to 0.6 of your current wealth}

\item{win}{percentage of wealth increase if you win
you win 1.5 of your current wealth}

\item{players}{number of players}

\item{rounds}{gamble rounds or repetitions}
}
\description{
!!!! Players and round both greater than 10^5 gives error
!!!! Error: cannot allocate vector of size 74.5 Gb
!!!! Limit players to <= 10^4 and rounds <= 10^4
!!!! Replicate Time spent on sapply -> lapply
}
