% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{coin_transform}
\alias{coin_transform}
\title{Adding Median, Mean and EV -> Transforming to long}
\usage{
coin_transform(x)
}
\arguments{
\item{x}{}
}
\description{
!!!! Biggest bottleneck is pivot longer !!!!
!!!! Consider keeping wide !!!!
}
